third_party/WebKit/LayoutTests/fast/css/font-face-multiple-ranges-for-unicode-range-expected.txt

unicode-range can have comma-separated multiple ranges. For each of the following pairs, the upper and the lower lines should have the same length on screen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

iiiiiiiiii test_0_0
iiiiiiiiii ref_0_0
jjjjjjjjjj test_0_1
jjjjjjjjjj ref_0_1
kkkkkkkkkk test_0_2
kkkkkkkkkk ref_0_2
llllllllll test_0_3
llllllllll ref_0_3

iiiiiiiiii test_1_0
iiiiiiiiii ref_1_0
jjjjjjjjjj test_1_1
jjjjjjjjjj ref_1_1
kkkkkkkkkk test_1_2
kkkkkkkkkk ref_1_2
llllllllll test_1_3
llllllllll ref_1_3

iiiiiiiiii test_2_0
iiiiiiiiii ref_2_0
jjjjjjjjjj test_2_1
jjjjjjjjjj ref_2_1
kkkkkkkkkk test_2_2
kkkkkkkkkk ref_2_2
llllllllll test_2_3
llllllllll ref_2_3

iiiiiiiiii test_3_0
iiiiiiiiii ref_3_0
jjjjjjjjjj test_3_1
jjjjjjjjjj ref_3_1
kkkkkkkkkk test_3_2
kkkkkkkkkk ref_3_2
llllllllll test_3_3
llllllllll ref_3_3

iiiiiiiiii test_4_0
iiiiiiiiii ref_4_0
jjjjjjjjjj test_4_1
jjjjjjjjjj ref_4_1
kkkkkkkkkk test_4_2
kkkkkkkkkk ref_4_2
llllllllll test_4_3
llllllllll ref_4_3

iiiiiiiiii test_5_0
iiiiiiiiii ref_5_0
jjjjjjjjjj test_5_1
jjjjjjjjjj ref_5_1
kkkkkkkkkk test_5_2
kkkkkkkkkk ref_5_2
llllllllll test_5_3
llllllllll ref_5_3

iiiiiiiiii test_6_0
iiiiiiiiii ref_6_0
jjjjjjjjjj test_6_1
jjjjjjjjjj ref_6_1
kkkkkkkkkk test_6_2
kkkkkkkkkk ref_6_2
llllllllll test_6_3
llllllllll ref_6_3


PASS document.getElementById('test_0_0').offsetWidth is document.getElementById('ref_0_0').offsetWidth
PASS document.getElementById('test_0_1').offsetWidth is document.getElementById('ref_0_1').offsetWidth
PASS document.getElementById('test_0_2').offsetWidth is document.getElementById('ref_0_2').offsetWidth
PASS document.getElementById('test_0_3').offsetWidth is document.getElementById('ref_0_3').offsetWidth
PASS document.getElementById('test_1_0').offsetWidth is document.getElementById('ref_1_0').offsetWidth
PASS document.getElementById('test_1_1').offsetWidth is document.getElementById('ref_1_1').offsetWidth
PASS document.getElementById('test_1_2').offsetWidth is document.getElementById('ref_1_2').offsetWidth
PASS document.getElementById('test_1_3').offsetWidth is document.getElementById('ref_1_3').offsetWidth
PASS document.getElementById('test_2_0').offsetWidth is document.getElementById('ref_2_0').offsetWidth
PASS document.getElementById('test_2_1').offsetWidth is document.getElementById('ref_2_1').offsetWidth
PASS document.getElementById('test_2_2').offsetWidth is document.getElementById('ref_2_2').offsetWidth
PASS document.getElementById('test_2_3').offsetWidth is document.getElementById('ref_2_3').offsetWidth
PASS document.getElementById('test_3_0').offsetWidth is document.getElementById('ref_3_0').offsetWidth
PASS document.getElementById('test_3_1').offsetWidth is document.getElementById('ref_3_1').offsetWidth
PASS document.getElementById('test_3_2').offsetWidth is document.getElementById('ref_3_2').offsetWidth
PASS document.getElementById('test_3_3').offsetWidth is document.getElementById('ref_3_3').offsetWidth
PASS document.getElementById('test_4_0').offsetWidth is document.getElementById('ref_4_0').offsetWidth
PASS document.getElementById('test_4_1').offsetWidth is document.getElementById('ref_4_1').offsetWidth
PASS document.getElementById('test_4_2').offsetWidth is document.getElementById('ref_4_2').offsetWidth
PASS document.getElementById('test_4_3').offsetWidth is document.getElementById('ref_4_3').offsetWidth
PASS document.getElementById('test_5_0').offsetWidth is document.getElementById('ref_5_0').offsetWidth
PASS document.getElementById('test_5_1').offsetWidth is document.getElementById('ref_5_1').offsetWidth
PASS document.getElementById('test_5_2').offsetWidth is document.getElementById('ref_5_2').offsetWidth
PASS document.getElementById('test_5_3').offsetWidth is document.getElementById('ref_5_3').offsetWidth
PASS document.getElementById('test_6_0').offsetWidth is document.getElementById('ref_6_0').offsetWidth
PASS document.getElementById('test_6_1').offsetWidth is document.getElementById('ref_6_1').offsetWidth
PASS document.getElementById('test_6_2').offsetWidth is document.getElementById('ref_6_2').offsetWidth
PASS document.getElementById('test_6_3').offsetWidth is document.getElementById('ref_6_3').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
